requirements.txt
alembic==1.0.8
bcrypt==3.1.6
Flask==1.0.2
Flask-Admin==1.5.3
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
Pillow==6.0.0
SQLAlchemy==1.3.2
WTForms==2.2.1
